# More efficient decoding of arrays