Tested with Scala 3 and JDK 21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfBigDecimalsReading.zioJson     128  thrpt    5   151187.739 ±   459.966  ops/s
ArrayOfBigIntsReading.zioJson         128  thrpt    5   137918.018 ±  1377.553  ops/s
ArrayOfBytesReading.zioJson           128  thrpt    5  1013394.269 ±  5615.047  ops/s
ArrayOfDoublesReading.zioJson         128  thrpt    5   304200.753 ±  4779.653  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   443325.617 ±  6852.218  ops/s
ArrayOfIntsReading.zioJson            128  thrpt    5   778311.952 ±  6004.551  ops/s
ArrayOfLongsReading.zioJson           128  thrpt    5   477036.224 ±  7342.575  ops/s
ArrayOfShortsReading.zioJson          128  thrpt    5  1044668.559 ± 35687.214  ops/s
```

#### After
```txt
Benchmark                          (size)   Mode  Cnt        Score        Error  Units
ArrayOfBigDecimalsReading.zioJson     128  thrpt    5   141654.805 ±   3124.563  ops/s
ArrayOfBigIntsReading.zioJson         128  thrpt    5   140655.497 ±   1978.970  ops/s
ArrayOfBytesReading.zioJson           128  thrpt    5  1282858.364 ±   7311.231  ops/s
ArrayOfDoublesReading.zioJson         128  thrpt    5   315393.767 ±   4103.181  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   460529.057 ±  13222.819  ops/s
ArrayOfIntsReading.zioJson            128  thrpt    5   907623.346 ±  15618.025  ops/s
ArrayOfLongsReading.zioJson           128  thrpt    5   490157.638 ±   7324.797  ops/s
ArrayOfShortsReading.zioJson          128  thrpt    5  1257591.173 ±  11435.255  ops/s
```